Fix #82 float parser

+++ cql-parser/src/main/scala/troy/cql/parser/CqlParser.scala
@@ -107,7 +107,7 @@ object CqlParser extends JavaTokenParsers
 
     def float = floatingPointNumber
 
-    def number = integer | float
+    def number = float | integer
 
     def uuid = "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}".r
